Add fastapi Project that is able to process workflow log messages
Nextflow has a switch that allows to generate log messages which can be send to a webserver. The FastAPI Projects is the server used for this purpose,
mainly because it was the easiest way to setup a working minimal example.
This part of the project is intended to analyze the messages send by nextflow as well as experiment with filtering and displaying them. It is most likely
not going to be used in production code.