Store messages and send previous messages to websocket on connect
In order to be able to restore the state of any graph/ progressbar/ whatever, it is vital to send all messages known to the server upon establishing the new connection.
Tested this behaviour with two connections and multiple reloads, seem to work fine.
This commit will be used to actually implement the progress graph/values