Add message service and various debug components
Next step is to get the weblog messages into the nuxt project. This is done by adding a global service that stores all messages. Later, filter functions and logic to
discard messages not intended for weblog can be added to the service.
Then, each component can only use this global store to access the currently available messages.
If we want to restore a status, we can load old messages into this store, too. (This is our single point of weblog truth)

Sending and receiving messages via the websocket is not tested yet.