Add websocket client for fastapi project
For testing purposes, this commit adds a new plugin that connects to the websocket provided by the FastAPI server. As of now, messages are only logged in the console. Pusing the data into the global store does not work as expected.

Note that the FastAPI project needs to be running on port 8765 in order to work.