Make message items observable
The component is not updated immediately as the array has not been reactive. By making it a vue.observable, any updates are passed down to the component.
This still is not a nice solution as using the object is rather tricky (thankfully, it is encapsulated in the ListView etc. so modifications should be doable)